fix(web): move percentage helper outside server action
